[Function] addCart added. [Optimize] More strict checks are added before arguments are passed in.

+++ elm-v2.0/src/main/java/cn/edu/tju/elm/repository/FoodRepository.java
@@ -4,8 +4,7 @@ import cn.edu.tju.elm.model.Food;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
-public interface FoodRepository extends JpaRepository<Food, Integer> {
+public interface FoodRepository extends JpaRepository<Food, Long> {
     List<Food> findAllByBusinessId(Long businessId);
 }
